Moved command_line.h from tools/ to config/ ; updated use of command line functions throughout.
