+++ UTests/scholar/Author.cc
@@ -1,17 +1,17 @@
-// This file is part of Empirical, https://github.com/mercere99/Empirical/, and is  
-// Copyright (C) Michigan State University, 2015. It is licensed                
-// under the MIT Software license; see doc/LICENSE
+//  This file is part of Empirical, https://github.com/devosoft/Empirical
+//  Copyright (C) Michigan State University, 2016.
+//  Released under the MIT Software license; see doc/LICENSE
 
 #include <iostream>
 
+#include "../../config/command_line.h"
 #include "../../tools/assert.h"
-#include "../../tools/command_line.h"
 #include "../../scholar/Author.h"
 
 int main(int argc, char* argv[])
 {
   std::vector<std::string> args = emp::cl::args_to_strings(argc, argv);
-  const bool verbose = emp::cl::use_arg(args, "-v");
+  const bool verbose = emp::cl::use_flag(args, "-v");
 
   emp::Author fml("First", "Middle", "Last");
   emp_assert(fml.GetFullName() == "First Middle Last");
